Implemented additional API for cart service.

+++ cart-service/src/main/java/com/capstone/ecommerce/cart/entity/CheckoutInfo.java
@@ -4,6 +4,6 @@ public record CheckoutInfo(
         String orderId,
         Long userId,
         Long addressId,
-        Long paymentMethodId,
+        int paymentMethodId,
         String paymentLink
 ) {}
